focus refactor into redux logic"
